# Speed up qp problem encoding