--- a/releasenotes/notes/speed-up-qp-encoding-f8ae1d86fa1dfde2.yaml
+++ b/releasenotes/notes/speed-up-qp-encoding-f8ae1d86fa1dfde2.yaml
@@ -1,0 +1,4 @@
+---
+features:
+  - |
+    Speed-up ``dwave.cloud.utils.qubo.active_qubits`` utility function by ~50%.