ASV benchmarks on my machine:
- `active_qubits()`: `34 ms -> 22 ms`
- `active_qubits()` on BQM: `156 ms -> 121ms`
- `encode_problem_as_qp()`: `103 ms -> 75 ms`
- `encode_problem_as_qp()` on BQM: `650 ms -> 560 ms` and then ` -> 400 ms` when converted to `dict` upfront